+++ ledger/participant-state/kvutils/tools/BUILD.bazel
@@ -10,12 +10,39 @@ load(
     "lf_scalacopts",
 )
 
-da_scala_library(
-    name = "tools",
-    srcs = glob(["src/main/scala/**/*.scala"]),
-    tags = ["maven_coordinates=com.daml:data-export-kvutils:__VERSION__"],
+da_scala_binary(
+    name = "integrity-check",
+    srcs = glob(["integrity-check-v1/src/main/scala/ledger/participant/state/kvutils/tools/integritycheck/v1/**/*.scala"]),
+    main_class = "com.daml.ledger.participant.state.kvutils.tools.integritycheck.v1.Main",
+    resources = glob(["src/main/resources/*"]),
     visibility = [
         "//visibility:public",
+    ],
+    runtime_deps = [
+        "@maven//:ch_qos_logback_logback_classic",
+    ],
+    deps = [
+        "//daml-lf/data",
+        "//daml-lf/engine",
+        "//ledger/metrics",
+        "//ledger/participant-state",
+        "//ledger/participant-state/kvutils",
+        "//ledger/participant-state/kvutils:daml_kvutils_java_proto",
+        "@maven//:com_google_protobuf_protobuf_java",
+        "@maven//:io_dropwizard_metrics_metrics_core",
+    ],
+)
+
+da_scala_binary(
+    name = "integrity-check-v2",
+    srcs = glob(["integrity-check-v2/src/main/scala/ledger/participant/state/kvutils/tools/integritycheck/v2/**/*.scala"]),
+    main_class = "com.daml.ledger.participant.state.kvutils.tools.integritycheck.v2.Main",
+    resources = glob(["src/main/resources/*"]),
+    visibility = [
+        "//visibility:public",
+    ],
+    runtime_deps = [
+        "@maven//:ch_qos_logback_logback_classic",
     ],
     deps = [
         "//daml-lf/data",
@@ -33,42 +60,12 @@ da_scala_library(
     ],
 )
 
-da_scala_binary(
-    name = "integrity-check",
-    main_class = "com.daml.ledger.participant.state.kvutils.tools.integritycheck.v1.Main",
-    resources = glob(["src/main/resources/*"]),
-    visibility = [
-        "//visibility:public",
-    ],
-    runtime_deps = [
-        "@maven//:ch_qos_logback_logback_classic",
-    ],
-    deps = [
-        ":tools",
-    ],
-)
-
-da_scala_binary(
-    name = "integrity-check-v2",
-    main_class = "com.daml.ledger.participant.state.kvutils.tools.integritycheck.v2.Main",
-    resources = glob(["src/main/resources/*"]),
-    visibility = [
-        "//visibility:public",
-    ],
-    runtime_deps = [
-        "@maven//:ch_qos_logback_logback_classic",
-    ],
-    deps = [
-        ":tools",
-    ],
-)
-
 da_scala_test(
-    name = "tools-tests",
+    name = "integrity-check-v2-tests",
     size = "small",
-    srcs = glob(["src/test/**/IntegrityCheckerSpec.scala"]),
+    srcs = glob(["integrity-check-v2/src/test/scala/ledger/participant/state/kvutils/tools/integritycheck/v2/IntegrityCheckerSpec.scala"]),
     deps = [
-        ":tools",
+        ":integrity-check-v2",
         "//ledger/participant-state/kvutils",
         "@maven//:com_google_protobuf_protobuf_java",
         "@maven//:org_mockito_mockito_core",
@@ -78,10 +75,7 @@ da_scala_test(
 
 da_scala_benchmark_jmh(
     name = "benchmark-replay",
-    srcs = glob([
-        "src/test/scala/**/Replay.scala",
-        "src/test/scala/**/Adapter.scala",
-    ]),
+    srcs = glob(["integrity-check-v2/src/benchmark/scala/**/*"]),
     max_heap_size = "4g",
     scalacopts = lf_scalacopts,
     deps = [
