+++ ledger/participant-state/kvutils/tools/integrity-check-v2/src/benchmark/scala/ledger/participant/state/kvutils/tools/integritycheck/v2/benchmark/Replay.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.kvutils.test
+package com.daml.ledger.participant.state.kvutils.tools.integritycheck.v2.benchmark
 
 import java.lang.System.err.println
 import java.nio.file.{Path, Paths}
