+++ ledger/participant-state/kvutils/tools/integrity-check-v2/src/main/scala/ledger/participant/state/kvutils/tools/integritycheck/v2/WriteRecordingLedgerStateOperations.scala
