+++ ledger/participant-state/kvutils/tools/integrity-check-v2/src/test/scala/ledger/participant/state/kvutils/tools/integritycheck/v2/IntegrityCheckerSpec.scala
